course ownership update and delete

+++ controllers/courses.js
@@ -37,12 +37,13 @@ exports.getCourse = asyncHandeler(async (req, res, next) => {
   res.status(200).json({ success: true, data: course });
 });
 
-//@DESC        Add a course
+//@DESC        Add course
 //@ROUTE       POST /api/v1/bootcamps/:bootcampsId/courses
 //@ACCESS      Private
 exports.addCourse = asyncHandeler(async (req, res, next) => {
-  // adds course to bootcamp
+  // got bootcamp id from params and put into req.boduy
   req.body.bootcamp = req.params.bootcampId;
+  req.body.user = req.user.id;
 
   const bootcamp = await Bootcamp.findById(req.params.bootcampId);
 
@@ -52,6 +53,17 @@ exports.addCourse = asyncHandeler(async (req, res, next) => {
       404
     );
   }
+
+  // Make sure bootcamp owner is the user logged in
+  if (bootcamp.user.toString() !== req.user.id && req.user.role !== 'admin') {
+    return next(
+      new ErrorResponse(
+        `User ${req.user.id} is not authorized to add a course to bootcamp ${bootcamp._id}`,
+        401
+      )
+    );
+  }
+
   const course = await Course.create(req.body);
 
   res.status(200).json({ success: true, data: course });
@@ -69,6 +81,17 @@ exports.updateCourse = asyncHandeler(async (req, res, next) => {
       404
     );
   }
+
+  // Make sure user is course owner
+  if (course.user.toString() !== req.user.id && req.user.role !== 'admin') {
+    return next(
+      new ErrorResponse(
+        `User ${req.user.id} is not authorized to update course ${course._id}`,
+        401
+      )
+    );
+  }
+
   course = await Course.findByIdAndUpdate(req.params.id, req.body, {
     new: true,
     runValidators: true,
@@ -89,6 +112,17 @@ exports.deleteCourse = asyncHandeler(async (req, res, next) => {
       404
     );
   }
+
+  // Make sure user is course owner
+  if (course.user.toString() !== req.user.id && req.user.role !== 'admin') {
+    return next(
+      new ErrorResponse(
+        `User ${req.user.id} is not authorized to delete course ${course._id}`,
+        401
+      )
+    );
+  }
+
   await course.deleteOne();
   res.status(200).json({ success: true, data: {} });
 });
